requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
bidict==0.21.4
python-dotenv==0.19.0
gevent==21.8.0
gevent-websocket==0.10.1
PyAudio==0.2.11
opencv-python==4.5.3.56
numpy==1.21.2
pillow==8.3.2